+++ effectful/CHANGELOG.md
@@ -1,6 +1,7 @@
 # effectful-2.2.2.0 (????-??-??)
 * Add `withSeqEffToIO` and `withConcEffToIO` to `Effectful`.
 * Use strict `IORef` and `MVar` variants where appropriate.
+* Add `reinterpret2` and `reinterpret3` to `Effectful.Dispatch.Dynamic`.
 
 # effectful-2.2.1.0 (2022-11-09)
 * Add `localSeqLift` and `localLift` to `Effectful.Dispatch.Dynamic`.